spike: CR Feedback; EngineAPIService convergence
